requirements.txt
aws-cdk-lib==2.136.1
constructs>=10.0.0,<11.0.0
requests
boto3